Added abstract base class for scenario events and separate event for scenario removal. Consolidated briefing tab scenario handlers into one handler for the base class.
