+++ MekHQ/src/mekhq/gui/BriefingTab.java
@@ -58,8 +58,7 @@ import mekhq.campaign.event.MissionChangedEvent;
 import mekhq.campaign.event.MissionNewEvent;
 import mekhq.campaign.event.OptionsChangedEvent;
 import mekhq.campaign.event.OrganizationChangedEvent;
-import mekhq.campaign.event.ScenarioChangedEvent;
-import mekhq.campaign.event.ScenarioNewEvent;
+import mekhq.campaign.event.ScenarioEvent;
 import mekhq.campaign.event.ScenarioResolvedEvent;
 import mekhq.campaign.force.Force;
 import mekhq.campaign.force.Lance;
@@ -970,19 +969,8 @@ public final class BriefingTab extends CampaignGuiTab {
     }
     
     @Subscribe
-    public void scenarioNew(ScenarioNewEvent ev) {
+    public void scenarioHandler(ScenarioEvent ev) {
         scenarioListScheduler.schedule();
-    }
-    
-    @Subscribe
-    public void scenarioChanged(ScenarioChangedEvent ev) {
-        scenarioListScheduler.schedule();
-    }
-    
-    @Subscribe
-    public void scenarioResolved(ScenarioChangedEvent ev) {
-        scenarioListScheduler.schedule();
-        changeMission();
     }
     
     @Subscribe
